+++ psycopg/psycopg/connection_async.py
@@ -138,7 +138,7 @@ class AsyncConnection(BaseConnection[Row]):
                 await self.rollback()
             except Exception as exc2:
                 logger.warning(
-                    "error ignored rolling back transaction on %s: %s",
+                    "error ignored in rollback on %s: %s",
                     self,
                     exc2,
                 )