+++ tests/test_connection_async.py
@@ -169,7 +169,7 @@ async def test_context_close(aconn):
         await aconn.close()
 
 
-async def test_context_rollback_no_clobber(conn, dsn, caplog):
+async def test_context_inerror_rollback_no_clobber(conn, dsn, caplog):
     with pytest.raises(ZeroDivisionError):
         async with await psycopg.AsyncConnection.connect(dsn) as conn2:
             await conn2.execute("select 1")
@@ -182,7 +182,25 @@ async def test_context_rollback_no_clobber(conn, dsn, caplog):
     assert len(caplog.records) == 1
     rec = caplog.records[0]
     assert rec.levelno == logging.WARNING
-    assert "rolling back" in rec.message
+    assert "in rollback" in rec.message
+
+
+async def test_context_active_rollback_no_clobber(conn, dsn, caplog):
+    caplog.set_level(logging.WARNING, logger="psycopg")
+
+    with pytest.raises(ZeroDivisionError):
+        async with await psycopg.AsyncConnection.connect(dsn) as conn2:
+            async with conn2.cursor() as cur:
+                async with cur.copy(
+                    "copy (select generate_series(1, 10)) to stdout"
+                ) as copy:
+                    async for row in copy.rows():
+                        1 / 0
+
+    assert len(caplog.records) == 1
+    rec = caplog.records[0]
+    assert rec.levelno == logging.WARNING
+    assert "in rollback" in rec.message
 
 
 @pytest.mark.slow