+++ tests/test_psycopg_dbapi20.py
@@ -6,6 +6,7 @@ import psycopg
 from psycopg.conninfo import conninfo_to_dict
 
 from . import dbapi20
+from . import dbapi20_tpc
 
 
 @pytest.fixture(scope="class")
@@ -28,8 +29,21 @@ class PsycopgTests(dbapi20.DatabaseAPI20Test):
         pass
 
 
+@pytest.mark.usefixtures("tpc")
+@pytest.mark.usefixtures("with_dsn")
+class PsycopgTPCTests(dbapi20_tpc.TwoPhaseCommitTests):
+    driver = psycopg
+    connect_args = ()  # set by the fixture
+
+    def connect(self):
+        return psycopg.connect(*self.connect_args)
+
+
 # Shut up warnings
 PsycopgTests.failUnless = PsycopgTests.assertTrue  # type: ignore[assignment]
+PsycopgTPCTests.assertEquals = (  # type: ignore[assignment]
+    PsycopgTPCTests.assertEqual
+)
 
 
 @pytest.mark.parametrize(